requirements.txt
flask~=3.1.1
flask-cors~=6.0.0

# API Framework
fastapi>=0.92.0
uvicorn>=0.20.0
python-multipart>=0.0.5
pydantic>=1.10.5

# Database
pymongo[srv]>=4.3.3

# Storage
google-cloud-storage>=2.7.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Environment variables
python-dotenv>=0.21.1

# HTTP Client (for ML service communication)
requests>=2.28.2

# Testing
pytest>=7.2.2
httpx>=0.23.3

# requirements.txt

pip freeze > requirements.txt

flask==2.0.1
gunicorn==20.1.0
pymongo==4.0.1
google-cloud-storage==2.1.0
numpy==1.22.3
scikit-learn==1.0.2
xgboost==1.5.1
librosa==0.9.1
soundfile==0.10.3.post1
flask-cors==3.0.10
python-dotenv==0.19.2
Werkzeug==2.0.2
joblib==1.1.0
pydub==0.25.1
mongomock>=4.1.2
pytest-cov>=4.1.0